PF/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
Django==4.1.3
django-recurrence==1.11.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
tzlocal==4.2
